Refactor user ID type to int64 and implement UserController
- Update UserRepository and UserUsecase interfaces to use int64 for user ID.
- Introduce UserController to handle HTTP requests related to user operations.
- Implement methods for retrieving all users and a user by ID in the UserController, including error handling.
- Add gin-gonic/gin v1.11.0 as a direct dependency.
- Include several indirect dependencies for JSON handling, validation, and other utilities.
- Create go.sum file to ensure consistent dependency resolution.

+++ internal/infrastructure/repository/user.go
@@ -8,7 +8,7 @@ import (
 
 type UserRepository interface {
 	SelectAll(ctx context.Context) ([]domain.User, error)
-	SelectByID(ctx context.Context, id int) (domain.User, error)
+	SelectByID(ctx context.Context, id int64) (domain.User, error)
 	SelectByEmail(ctx context.Context, email string) (domain.User, error)
 	SelectByUsername(ctx context.Context, username string) (domain.User, error)
 	Insert(ctx context.Context, user domain.User) (domain.User, error)
@@ -44,7 +44,7 @@ func (u User) SelectAll(ctx context.Context) ([]domain.User, error) {
 	return users, nil
 }
 
-func (u User) SelectByID(ctx context.Context, id int) (domain.User, error) {
+func (u User) SelectByID(ctx context.Context, id int64) (domain.User, error) {
 	row := u.db.QueryRowContext(ctx, "SELECT id, username, email, password, created_at, updated_at FROM users WHERE id = $1", id)
 	var user domain.User
 	err := row.Scan(&user.ID, &user.Username, &user.Email, &user.Password, &user.CreatedAt, &user.UpdatedAt)
